+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Change
 
+  - When location is set in map-response, put it in response json also
   - Module component - image update action 
   - Deployment - component image update action
   - The user resource ACL grants the user all rights except edit-acl.